VTK-ify

+++ vtkPointSetNormalOrientationToPoint.cxx
@@ -17,9 +17,9 @@ vtkStandardNewMacro(vtkPointSetNormalOrientationToPoint);
 
 vtkPointSetNormalOrientationToPoint::vtkPointSetNormalOrientationToPoint()
 {
-  this->OrientationPoint[0] = 0.0;
-  this->OrientationPoint[1] = 0.0;
-  this->OrientationPoint[2] = 0.0;
+  this->OrientationPoint[0] = static_cast<double>(0);
+  this->OrientationPoint[1] = static_cast<double>(0);
+  this->OrientationPoint[2] = static_cast<double>(0);
 }
 
 int vtkPointSetNormalOrientationToPoint::RequestData(vtkInformation *vtkNotUsed(request),
